Improved testing efforts
Tests for lpm init and lpm install --lockfile

+++ src/helpers/FileHelper.py
@@ -11,5 +11,8 @@ def clear_and_remove_packages_folder():
     pkg_dir = config.get_package_dir()
     path = os.path.abspath(pkg_dir)
 
-    logger.info(f"Deleting all contents of {path}")
-    shutil.rmtree(path)
+    if os.path.exists(path):
+        logger.info(f"Deleting all contents of {path}")
+        shutil.rmtree(path)
+    else:
+        logger.debug(path + ' already removed')
